警告されなくなったよ２

+++ player/player/object.cpp
@@ -122,7 +122,7 @@ void CObject::set_layer_id(DRAW_LAYER layer_id_){
 
 //size:文字サイズ0~1のfloat space:文字間の間隔 int ,str:文字列 string
 void CObject::DrawOriginalString(int x,int y,float size,int space,string str){
-	for(int i = 0; i < str.length(); i++){
-		DrawExtendGraph(x+space*i,y,x+space*i+49*size,y+51*size,word[(int)str[i]],true);
+	for(int i = 0; i < (int)str.length(); i++){
+		DrawExtendGraph(x+space*i,y,x+space*i+(int)(49*size),y+(int)(51*size),word[(int)str[i]],true);
 	}
 }
